Fix error when calling :T
The `:T` command called without first calling `Open` throws errors
instead of sending the output to the terminal. This is because the
terminal job ids are not setup yet.

Resolve this by calling `Setup` when the terminal is opened, and if
there is no job id yet, open a new terminal in a split.